+++ src/popup/pages/CredentialManage/index.js
@@ -1,6 +1,6 @@
 import { getCredentialDisplayData } from "@/utils/utils";
+import { getSimplifyCredentialData } from "@/utils/o1jsUtils";
 import i18n from "i18next";
-import { PrettyPrinter } from "mina-attestations";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
@@ -51,9 +51,7 @@ const CredentialManage = ({}) => {
         let list = [];
         for (let index = 0; index < credentials.length; index++) {
           const credential = credentials[index];
-          const displayCredentialData = PrettyPrinter.simplifyCredentialData(
-            JSON.parse(credential.credential)
-          );
+          const displayCredentialData = getSimplifyCredentialData(JSON.parse(credential.credential));
           const parseCredential = getCredentialDisplayData(
             displayCredentialData
           );