+++ src/utils/o1jsUtils.firefox.js
@@ -1,0 +1,5 @@
+export const getSimplifyCredentialData = () => ({});
+export const getPrintPresentationRequest = () => "";
+export const getPrintVerifierIdentity = () => "";
+export const checkPresentationRequestSchema = () => ({ success: true });
+export const checkStoredCredentialSchema = () => ({ success: true });